+++ prettier.config.mjs
@@ -1,0 +1,8 @@
+/** @type {import('prettier').Config} */
+const config = {
+  printWidth: 120,
+  singleQuote: true,
+  trailingComma: 'all',
+};
+
+export default config;